refactor(new): standardize project name validation
